fix: widen restaurant ID type from u8 to u32

+++ src/search.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 pub struct Restaurant {
     #[serde(rename = "openingHours")]
     opening_hours: Vec<Option<String>>,
-    pub id: u8,
+    pub id: u32,
     name: String,
     url: String,
     address: String,
@@ -15,7 +15,7 @@ pub struct Restaurant {
 impl Restaurant {
     #[cfg(test)]
     pub fn new(
-        id: u8,
+        id: u32,
         name: String,
         url: String,
         address: String,
